+++ src/main/java/com/github/games647/lagmonitor/commands/HeapCommand.java
@@ -4,13 +4,18 @@ import com.github.games647.lagmonitor.LagMonitor;
 import com.github.games647.lagmonitor.Pagination;
 import com.google.common.collect.Lists;
 
+import java.io.File;
 import java.lang.management.ManagementFactory;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
 import java.util.logging.Level;
 
 import javax.management.MBeanServer;
 import javax.management.ObjectName;
 
+import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.chat.BaseComponent;
 import net.md_5.bungee.api.chat.ComponentBuilder;
 
@@ -22,15 +27,19 @@ import org.bukkit.command.CommandSender;
 public class HeapCommand implements CommandExecutor {
 
     //https://docs.oracle.com/javase/8/docs/jre/api/management/extension/com/sun/management/DiagnosticCommandMBean.html
-    private static final String DIAGNOTISC_COMMAND = "com.sun.management:type=DiagnosticCommand";
-
+    private static final String DIAGNOSTIC_COMMAND = "com.sun.management:type=DiagnosticCommand";
     private static final String HEAP_COMMAND = "gcClassHistogram";
 
     //can be useful for dumping heaps in binary format
     //https://docs.oracle.com/javase/8/docs/jre/api/management/extension/com/sun/management/HotSpotDiagnosticMXBean.html
     private static final String HOTSPOT_DIAGNOSTIC = "com.sun.management:type=HotSpotDiagnostic";
+    private static final String DUMP_COMMAND = "dumpHeap";
+    private static final String DUMP_FILE_NAME = "heap";
+    private static final String DUMP_FILE_ENDING = ".hprof";
+    private static final boolean DUMP_DEAD_OBJECTS = false;
 
     private final LagMonitor plugin;
+    private final DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd-HH-mm-ss");
 
     public HeapCommand(LagMonitor plugin) {
         this.plugin = plugin;
@@ -38,12 +47,22 @@ public class HeapCommand implements CommandExecutor {
 
     @Override
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        MBeanServer mBeanServer = ManagementFactory.getPlatformMBeanServer();
+        if (args.length > 0) {
+            String subCommand = args[0];
+            if ("dump".equalsIgnoreCase(subCommand)) {
+                onDump(sender);
+            } else {
+                sender.sendMessage(ChatColor.DARK_RED + "Unknown subcommand");
+            }
+
+            return true;
+        }
 
         List<BaseComponent[]> paginatedLines = Lists.newArrayList();
         try {
-            ObjectName diagnosticObjectName = ObjectName.getInstance(DIAGNOTISC_COMMAND);
+            ObjectName diagnosticObjectName = ObjectName.getInstance(DIAGNOSTIC_COMMAND);
 
+            MBeanServer mBeanServer = ManagementFactory.getPlatformMBeanServer();
             String reply = (String) mBeanServer.invoke(diagnosticObjectName, HEAP_COMMAND
                     , new Object[]{ArrayUtils.EMPTY_STRING_ARRAY}, new String[]{String[].class.getName()});
             String[] lines = reply.split("\n");
@@ -51,13 +70,34 @@ public class HeapCommand implements CommandExecutor {
                 paginatedLines.add(new ComponentBuilder(line).create());
             }
 
+            Pagination pagination = new Pagination("Heap", paginatedLines);
+            pagination.send(sender);
+            plugin.getPaginations().put(sender, pagination);
         } catch (Exception ex) {
             plugin.getLogger().log(Level.SEVERE, null, ex);
+            sender.sendMessage(ChatColor.DARK_RED + "An exception occurred. Please check the server log");
         }
 
-        Pagination pagination = new Pagination("Heap", paginatedLines);
-        pagination.send(sender);
-        plugin.getPaginations().put(sender, pagination);
         return true;
+    }
+
+    private void onDump(CommandSender sender) {
+        try {
+            MBeanServer mBeanServer = ManagementFactory.getPlatformMBeanServer();
+            ObjectName hotspotBean = ObjectName.getInstance(HOTSPOT_DIAGNOSTIC);
+
+            String timeSuffix = '-' + dateFormat.format(new Date());
+            File dumpFile = new File(plugin.getDataFolder(), DUMP_FILE_NAME + timeSuffix + DUMP_FILE_ENDING);
+            //it needs to be with a system dependent path seperator
+            mBeanServer.invoke(hotspotBean, DUMP_COMMAND
+                    , new Object[]{dumpFile.getAbsolutePath(), DUMP_DEAD_OBJECTS}
+                    , new String[]{String.class.getName(), Boolean.TYPE.getName()});
+
+            sender.sendMessage(ChatColor.GRAY + "Dump created: " + dumpFile.getCanonicalPath());
+            sender.sendMessage(ChatColor.GRAY + "You can analyse it using VisualVM");
+        } catch (Exception ex) {
+            plugin.getLogger().log(Level.SEVERE, null, ex);
+            sender.sendMessage(ChatColor.DARK_RED + "An exception occurred. Please check the server log");
+        }
     }
 }